# docs/static/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: eInvoice Storage - OpenAPI 3.0
  description: |-
    A description of the Invoicer REST API
  contact:
    email: patricklachlanbell@gmail.com
  version: 1.0.1
...
servers:
  - url: http://node-express-env.eba-9atmj7qr.ap-southeast-2.elasticbeanstalk.com
    description: Production server (uses live data and can be viewed on frontend)
paths:
...
  /invoices/feed:
    get:
      tags:
        - invoices
      summary: Gets a page of the users invoices.
      description: Paginated feed of the the invoices a user has permission to view. The feed is in sorted from newest to oldest (in future there will be a sort parameter).
      parameters:
        - in: query
          name: page
          description: Zero indexed number representing which page of results to return, each page is of length 'limit.' The first invoice is invoice number 'page' x 'limit.'
          schema:
            type: number
            minimum: 0
            example: 0
        - in: query
          name: offset
          description: The number of invoices to offset the page by, this does not affect how many invoices are returned.
          schema:
            type: number
            minimum: 0
            example: 0
        - in: query
          name: limit
          description: The size of each page, i.e. the maximum number of invoices which are returned.
          schema:
            type: number
            default: 10
            maximum: 100
            minimum: 1
            example: 20
      responses:
        '200':
          description: An object respresenting the specified page of invoices and how many invoices are left in the feed. There can be less than 'limit' invoices in 'page'.
          content:
            application/json:
              schema:
                type: object
                required:
                  - page
                  - total
                properties:
                  page:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  total:
                    type: integer
                    description: The total number of invoices in the feed.
                    example: 100
        '400':
          $ref: '#/components/responses/InvalidError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
      security:
        - bearerAuth:
          - view:invoice
  /invoices/search:
    get:
      tags:
        - invoices
      summary: Gets a page of invoices which match the provided query.
      description: Paginated results from a fuzzy search from the invoices the user has permission to view. The results are sorted from newest to oldest (in future there will be a sort parameter).
      parameters:
        - in: query
          name: query
          description: A fuzzy search term which determines which invoices are returned.
          schema:
            type: string
            example: Company A LLC.
        - in: query
          name: page
          description: Zero indexed number representing which page of results to return, each page is of length 'limit.' The first invoice is invoice number 'page' x 'limit.'
          schema:
            type: number
            minimum: 0
            example: 0
        - in: query
          name: offset
          description: The number of invoices to offset the page by, this does not affect how many invoices are returned.
          schema:
            type: number
            minimum: 0
            example: 0
        - in: query
          name: limit
          description: The size of each page, i.e. the maximum number of invoices which are returned.
          schema:
            type: number
            default: 10
            maximum: 100
            minimum: 1
            example: 20
      responses:
        '200':
          description: Paginated invoices which match with the query and are viewable by the authorizing user.
          content:
            application/json:
              schema:
                type: object
                required:
                  - page
                  - total
                properties:
                  page:
                    $ref: '#/components/schemas/Invoice'
                  total:
                    type: integer
                    description: The total number of invoices matching the query which the authorizing user can view.
                    example: 100
        '400':
          $ref: '#/components/responses/InvalidError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
  /invoices/batch:
    post:
      tags:
        - invoices
      summary: Applies an operation to a batch of invoices.
      description: Given a list of invoice ids and an operation, this operation is performed on all invoices which the authorizing user has edit permissions for. Currently the only supported operation is deletion.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
                - operation
              properties:
                ids:
                  type: array
                  description: Array of the objectIds of the invoices to perform the operation on.
                  example: ["6421d1e887d1e2174b494492"]
                  items:
                    type: string
                    example: "6421d1e887d1e2174b494492"
                operation:
                  type: string
                  description: A string reprenting the operation to perform.
                  enum:
                    - "delete"
                  example: "delete"
        required: true
        
      responses:
        '200':
          description: Acknowledgement and how many invoices the operation was applied to.
          content:
            application/json:
              schema:
                type: object
                required:
                  - acknowledged
                properties:
                  aknowledged:
                    type: boolean
                    example: true
                  deletedCount:
                    type: integer
                    description: If the operation is 'delete' then a delete count is returned. This is tells the caller how many of 'ids' were valid.
                    example: 1
        '400':
          $ref: '#/components/responses/InvalidError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Invoice:
      type: object
      required:
        - documentTitle
        - totalInvoiceValue
        - supplierName
        - recipientName
        - viewIds
        - editIds
        - creationTime
        - creatorId
      properties:
        contentsXml:
          example: "<Invoice></Invoice>"
          type: string
          minLength: 1
        contentsJson:
          type: string
          minLength: 1
        documentTitle:
          example: "Example"
          type: string
        totalInvoiceValue:
          example: 1000
          description: Currency is assumed to be AUD.
          type: number
        supplierName:
          example: "Company A LLC"
          type: string
        recipientName:
          example: "Company B LLC"
          type: string
        viewIds:
          example: ["6421d1e887d1e2174b494492"]
          type: array
          items:
            type: string
        editIds:
          example: ["6421d1e887d1e2174b494492"]
          type: array
          items:
            type: string
        creationTime:
          example: "1975-08-19T23:15:30.000Z"
          description: Serialized JS Date object, see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toJSON
          type: string
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, possibly because it has expired.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    InvalidError:
      description: Parameters don't match the schema (validated with [YUP](https://www.npmjs.com/package/yup)) or an internal server error was encountered.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: An optional message describing the error
                example: "field body.recipientName missing"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT